move client initialization to main
