+++ managed/cmd/pmm-managed/main.go
@@ -26,7 +26,7 @@ import (
 	"math/rand"
 	"net"
 	"net/http"
-	_ "net/http/pprof" //nolint:gosec // register /debug/pprof
+	_ "net/http/pprof"
 	"net/url"
 	"os"
 	"os/signal"
@@ -42,6 +42,8 @@ import (
 	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
 	grpc_gateway "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/pkg/errors"
+	metrics "github.com/prometheus/client_golang/api"
+	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	prom "github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
@@ -121,6 +123,9 @@ const (
 	defaultContextTimeout = 10 * time.Second
 	pProfProfileDuration  = 30 * time.Second
 	pProfTraceDuration    = 10 * time.Second
+
+	clickhouseMaxIdleConns = 5
+	clickhouseMaxOpenConns = 10
 )
 
 var pprofSemaphore = semaphore.NewWeighted(1)
@@ -632,6 +637,20 @@ func migrateDB(ctx context.Context, sqlDB *sql.DB, dbName, dbAddress, dbUsername
 	}
 }
 
+// newClickhouseDB return a new Clickhouse db.
+func newClickhouseDB(dsn string, maxIdleConns, maxOpenConns int) (*sql.DB, error) {
+	db, err := sql.Open("clickhouse", dsn)
+	if err != nil {
+		return nil, errors.Wrap(err, "Failed to open connection to QAN DB")
+	}
+
+	db.SetConnMaxLifetime(0)
+	db.SetMaxIdleConns(maxIdleConns)
+	db.SetMaxOpenConns(maxOpenConns)
+
+	return db, nil
+}
+
 func main() {
 	rand.Seed(time.Now().UnixNano())
 
@@ -712,8 +731,8 @@ func main() {
 
 	clickhouseDSN := "tcp://" + *clickhouseAddrF + "/" + *clickHouseDatabaseF
 
-	clickhouseDB := ds.QanDBSelect
-	clickhouseDB.DSN = clickhouseDSN
+	qanDB := ds.QanDBSelect
+	qanDB.DSN = clickhouseDSN
 
 	sqlDB, err := models.OpenDB(*postgresAddrF, *postgresDBNameF, *postgresDBUsernameF, *postgresDBPasswordF)
 	if err != nil {
@@ -798,7 +817,17 @@ func main() {
 
 	actionsService := agents.NewActionsService(qanClient, agentsRegistry)
 
-	checksService, err := checks.New(db, platformClient, actionsService, alertManager, *victoriaMetricsURLF, clickhouseDSN)
+	vmClient, err := metrics.NewClient(metrics.Config{Address: *victoriaMetricsURLF})
+	if err != nil {
+		l.Fatalf("Could not create Victoria Metrics client: %s", err)
+	}
+
+	clickhouseClient, err := newClickhouseDB(clickhouseDSN, clickhouseMaxIdleConns, clickhouseMaxIdleConns)
+	if err != nil {
+		l.Fatalf("Could not create Clickhouse client: %s", err)
+	}
+
+	checksService, err := checks.New(db, platformClient, actionsService, alertManager, v1.NewAPI(vmClient), clickhouseClient)
 	if err != nil {
 		l.Fatalf("Could not create checks service: %s", err)
 	}
